v4.34w.04.12.01

+++ src/extudp_functions.h
@@ -8,7 +8,7 @@
 void startExternUDP();
 void getExternUDP();
 void getExtern(unsigned char icomming[255], int len);
-void sendExtern(bool bUDP, char *src_type, uint8_t buffer[500], uint8_t buflen);
+void sendExtern(bool bUDP, char *src_type, uint8_t buffer[500], uint16_t buflen);
 void sendExternHeartbeat();
 void resetExternUDP();
